+++ activesupport/lib/active_support/core_ext/symbol/inquiry.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require "active_support/core_ext/string/inquiry"
+
+class Symbol
+  # See String#inquiry.
+  delegate :inquiry, to: :name
+end